requirements.txt
celery==5.3.4
Django==4.2.4
django-celery-results==2.5.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
redis==5.0.0
